Add "Why StoryBoard" article
This article focuses on explaining why we are changing from
Launchpad in the first place. It can be followed by articles
explaining how StoryBoard preserves most of the Launchpad
experience (Stories = Bugs), and another to explain how it
significantly differs from Launchpad (use of tasklists/boards
instead of priorities and milestones).